+++ data/processors/exports/ranking_factors.py
@@ -11,10 +11,10 @@ def export_ranking_factors_parquet(df: pl.DataFrame) -> None:
     OUTPUT_DIR.mkdir(exist_ok=True)
     extracted = df.select(
         pl.col("id"),
-        pl.col("ranking_rank_type").cast(pl.Int32).alias("rank_type"),
-        pl.col("ranking_rank_combined").cast(pl.Int32).alias("rank_combined"),
-        pl.col("ranking_rank_usage").cast(pl.Int32).alias("rank_usage"),
-        pl.col("ranking_rank_custom").cast(pl.Int32).alias("rank_custom"),
-        pl.col("ranking_rank_boost").cast(pl.Int32).alias("rank_boost"),
+        pl.col("ranking_rank_type").cast(pl.Int16).alias("rank_type"),
+        pl.col("ranking_rank_combined").cast(pl.Int16).alias("rank_combined"),
+        pl.col("ranking_rank_usage").cast(pl.Int16).alias("rank_usage"),
+        pl.col("ranking_rank_custom").cast(pl.Int16).alias("rank_custom"),
+        pl.col("ranking_rank_boost").cast(pl.Int16).alias("rank_boost"),
     ).unique()
     RankingFactorsSchema.write_parquet(extracted, OUTPUT_DIR / "ranking_factors.parquet")
